# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Analysis Canvas API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/requests/{request_id}/load-cases": {
      "get": {
        "summary": "Get Load Cases",
        "operationId": "get_load_cases_api_requests__request_id__load_cases_get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Load Cases Api Requests  Request Id  Load Cases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Load Case",
        "operationId": "create_load_case_api_requests__request_id__load_cases_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadCaseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Load Case Api Requests  Request Id  Load Cases Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/report-layouts": {
      "get": {
        "summary": "List Report Layouts",
        "operationId": "list_report_layouts_api_report_layouts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReportLayoutCatalogResponse"
                  },
                  "type": "array",
                  "title": "Response List Report Layouts Api Report Layouts Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Report Layout",
        "operationId": "create_report_layout_api_report_layouts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportLayoutPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportLayoutSavedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/report-layouts/{layout_id}": {
      "put": {
        "summary": "Update Report Layout",
        "operationId": "update_report_layout_api_report_layouts__layout_id__put",
        "parameters": [
          {
            "name": "layout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layout Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportLayoutPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportLayoutSavedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Report Layout",
        "operationId": "delete_report_layout_api_report_layouts__layout_id__delete",
        "parameters": [
          {
            "name": "layout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layout Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportLayoutDeactivationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/report-layouts/{layout_id}/versions": {
      "get": {
        "summary": "Get Report Layout Versions",
        "operationId": "get_report_layout_versions_api_report_layouts__layout_id__versions_get",
        "parameters": [
          {
            "name": "layout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layout Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReportLayoutVersionSummaryResponse"
                  },
                  "title": "Response Get Report Layout Versions Api Report Layouts  Layout Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/report-layouts/{layout_id}/versions/{version}": {
      "get": {
        "summary": "Get Report Layout Version",
        "operationId": "get_report_layout_version_api_report_layouts__layout_id__versions__version__get",
        "parameters": [
          {
            "name": "layout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layout Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportLayoutVersionDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoadCaseCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Name"
          },
          "analysis_type": {
            "type": "string",
            "enum": [
              "DROP",
              "SIDE_CLAMP"
            ],
            "title": "Analysis Type"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          }
        },
        "type": "object",
        "required": [
          "name",
          "analysis_type"
        ],
        "title": "LoadCaseCreate"
      },
...
      "ReportLayoutCatalogResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "definition": {
            "$ref": "#/components/schemas/ReportLayoutDefinition"
          },
          "is_system": {
            "type": "boolean",
            "title": "Is System"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "updated_by": {
            "type": "string",
            "title": "Updated By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "version",
          "definition",
          "is_system",
          "is_active",
          "created_at",
          "updated_at",
          "updated_by"
        ],
        "title": "ReportLayoutCatalogResponse"
      },
      "ReportLayoutDeactivationResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "deactivated",
            "title": "Status"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "id"
        ],
        "title": "ReportLayoutDeactivationResponse"
      },
      "ReportLayoutDefinition": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "coverVariant": {
            "type": "string",
            "enum": [
              "balanced",
              "executive",
              "evidence"
            ],
            "title": "Covervariant"
          },
          "accentColor": {
            "type": "string",
            "title": "Accentcolor"
          },
          "sectionOrder": {
            "items": {
              "type": "string",
              "enum": [
                "series",
                "scalar",
                "media"
              ]
            },
            "type": "array",
            "title": "Sectionorder"
          },
          "variablePlacements": {
            "items": {
              "$ref": "#/components/schemas/ReportVariablePlacement"
            },
            "type": "array",
            "title": "Variableplacements"
          },
          "includeMedia": {
            "type": "boolean",
            "title": "Includemedia"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "version",
          "coverVariant",
          "accentColor",
          "sectionOrder",
          "variablePlacements",
          "includeMedia"
        ],
        "title": "ReportLayoutDefinition",
        "description": "Stored layout JSON remains an extensible document contract."
      },
      "ReportLayoutPayload": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "title": "Description",
            "default": ""
          },
          "definition": {
            "additionalProperties": true,
            "type": "object",
            "title": "Definition"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          }
        },
        "type": "object",
        "required": [
          "name",
          "definition"
        ],
        "title": "ReportLayoutPayload"
      },
      "ReportLayoutSavedResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "definition": {
            "$ref": "#/components/schemas/ReportLayoutDefinition"
          },
          "is_system": {
            "type": "boolean",
            "title": "Is System"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "updated_by": {
            "type": "string",
            "title": "Updated By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "version",
          "definition",
          "is_system",
          "updated_at",
          "updated_by"
        ],
        "title": "ReportLayoutSavedResponse"
      },
      "ReportLayoutVersionDetailResponse": {
        "properties": {
          "layout_id": {
            "type": "string",
            "title": "Layout Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "definition": {
            "$ref": "#/components/schemas/ReportLayoutDefinition"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "layout_id",
          "version",
          "definition",
          "created_by",
          "created_at"
        ],
        "title": "ReportLayoutVersionDetailResponse"
      },
      "ReportLayoutVersionSummaryResponse": {
        "properties": {
          "layout_id": {
            "type": "string",
            "title": "Layout Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "is_valid": {
            "type": "boolean",
            "title": "Is Valid"
          }
        },
        "type": "object",
        "required": [
          "layout_id",
          "version",
          "created_by",
          "created_at",
          "is_valid"
        ],
        "title": "ReportLayoutVersionSummaryResponse"
      },
...
      "ReportVariablePlacement": {
        "properties": {
          "variableKey": {
            "type": "string",
            "title": "Variablekey"
          },
          "presentation": {
            "type": "string",
            "enum": [
              "chart",
              "table",
              "both"
            ],
            "title": "Presentation"
          },
          "order": {
            "type": "integer",
            "title": "Order"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "variableKey",
          "presentation",
          "order"
        ],
        "title": "ReportVariablePlacement",
        "description": "A stable variable placement while retaining layout-specific extensions."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}